# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description of my API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/create_category": {
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Create a new category",
                "description": "Creates a new category with the provided details.",
                "operationId": "a090d5ec28e5de32482531992444b335",
                "requestBody": {
                    "description": "Category details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "description": "Category photo",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "name_ru": {
                                        "description": "Name of the category in Russian",
                                        "type": "string"
                                    },
                                    "name_en": {
                                        "description": "Name of the category in English",
                                        "type": "string"
                                    },
                                    "name_fr": {
                                        "description": "Name of the category in French",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Category created successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/single_page_category": {
            "get": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Get information of a single category",
                "description": "Retrieves information of a single category based on the provided category_id.",
                "operationId": "de9753536ba7d86a68449806b0525cc4",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID of the category to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Category not found",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/update_category": {
            "post": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Update a category",
                "description": "Updates an existing category with the provided details.",
                "operationId": "5137db519eae626dd3cd21e94639bc26",
                "requestBody": {
                    "description": "Category details",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "ID of the category to update",
                                        "type": "integer"
                                    },
                                    "photo": {
                                        "description": "Category photo",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "name_ru": {
                                        "description": "Updated name of the category in Russian",
                                        "type": "string"
                                    },
                                    "name_en": {
                                        "description": "Updated name of the category in English",
                                        "type": "string"
                                    },
                                    "name_fr": {
                                        "description": "Updated name of the category in French",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Category updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Category updated successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/delete_category": {
            "delete": {
                "tags": [
                    "Admin Categories"
                ],
                "summary": "Delete a category",
                "description": "Deletes an existing category based on the provided category_id.",
                "operationId": "8fc6bb95c8db5d7ff0db9d9ac77e1231",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Category deleted successfully",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}